% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioc2016-pkg.R
\docType{data}
\name{dds_microglia}
\alias{dds_microglia}
\title{RNA-seq from microglia cells}
\description{
The \code{\link{SummarizedExperiment}} object for Bruttger et al, Immunity 2015. Three condition were assessed, namely bone marrow macrophages, wild type microglia, and repopulating microglia (GEO accession: GSE68376)
}
\details{
Read counts per gene for the RNA-seq experiment experiment performed in PMID: 26163371.
Interleukin-1 signaling is involved in local self-renewal and maintenance of microglia.
During early embryogenesis microglia arise from yolk sac progenitors populating the developing CNS, where they are maintained as tissue-resident macrophages throughout the organism’s lifespan.
Here, we describe an experimental system that allows the specific conditional ablation of microglia in vivo.
Strikingly, we found that the microglia compartment was reconstituted within one week following depletion. Microglia repopulation relied entirely on a CNS-resident, internal pool and was independent from bone marrow-derived precursors.
Newly formed microglia were found in highly proliferative, organized micro-clusters that dissolve once steady state was achieved. Gene expression profiling revealed prominent expression of Interleukin-1 (IL-1) receptor in proliferating microglia.
During the repopulation phase, IL-1 signaling was neutralized by treatment with IL-1 receptor antagonist that impaired microglia proliferation. Hence, microglia harbor a highly efficient potential to restore themselves without contribution of peripheral myeloid cells.
IL-1 signaling significantly participates in this restorative proliferation process and is involved in stabilizing microglia maintenance.
}
\author{
Federico Marini (\email{marinif@uni-mainz.de}),
Julia Bruttger, 2016
}
\references{
Bruttger J, Karram K, Woertge S, Regen T, Marini F, Hoppmann N, Klein M, Blank T, Yona S, Wolf Y, Mack M, Pinteaux E, Mueller W, Zipp F, Binder H, Bopp T, Prinz M, Jung S, Waisman A. Genetic Cell Ablation Reveals Clusters of Local Self-Renewing Microglia in the Mammalian Central Nervous System. Immunity. 2015 Jul 21;43(1):92-106. doi: 10.1016/j.immuni.2015.06.012. Epub 2015 Jul 7. PMID: 26163371. GEO: GSE68376
}

