% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioc2016-pkg.R
\docType{data}
\name{mg400bylocation}
\alias{mg400bylocation}
\title{A subset of a list of microglia signature genes, split by location/function}
\description{
A subset of a list of microglia signature genes, split by location/function, extracted from Butovsky et al (2014). Housekeeping genes were left out.
}
\author{
Federico Marini (\email{marinif@uni-mainz.de})
}
\references{
Oleg Butovsky, Mark P Jedrychowski,	Craig S Moore, Ron Cialic,	Amanda J Lanser, Galina Gabriely,	Thomas Koeglsperger,	Ben Dake,	Pauline M Wu,	Camille E Doykan,	Zain Fanek,	LiPing Liu,	Zhuoxun Chen,	Jeffrey D Rothstein, Richard M Ransohoff, Steven P Gygi,	Jack P Antel, Howard L Weiner, Identification of a unique TGF-beta-dependent molecular and functional signature in microglia, Nature Neuroscience 17, 131-143 (2014)
}

