% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioc2016-pkg.R
\docType{data}
\name{mg400signatureGenes}
\alias{mg400signatureGenes}
\title{A list of microglia signature genes}
\description{
A curated list of microglia signature genes, extracted from Butovsky et al (2014). Housekeeping genes were left out.
This list is represented on the MG400 microglial chip
}
\author{
Federico Marini (\email{marinif@uni-mainz.de})
}
\references{
Oleg Butovsky, Mark P Jedrychowski,	Craig S Moore, Ron Cialic,	Amanda J Lanser, Galina Gabriely,	Thomas Koeglsperger,	Ben Dake,	Pauline M Wu,	Camille E Doykan,	Zain Fanek,	LiPing Liu,	Zhuoxun Chen,	Jeffrey D Rothstein, Richard M Ransohoff, Steven P Gygi,	Jack P Antel, Howard L Weiner, Identification of a unique TGF-beta-dependent molecular and functional signature in microglia, Nature Neuroscience 17, 131-143 (2014)
}

