% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioc2016-pkg.R
\docType{package}
\name{rbioc2016}
\alias{rbioc2016}
\alias{rbioc2016-package}
\title{rbioc2016 - datasets used in the course and other extra material}
\description{
rbioc2016 contains the following datasets
}
\author{
Federico Marini \email{marinif@uni-mainz.de}, 2016

Maintainer: Federico Marini \email{marinif@uni-mainz.de}
}

