% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_libs.R
\name{setup_day1}
\alias{setup_day1}
\title{setup all packages for day 1 of the course}
\usage{
setup_day1()
}
\value{
Installs the required packages - day1
}
\description{
setup all packages for day 1 of the course
}
\examples{
\dontrun{
 setup_day1()
}
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2016
}

