% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_libs.R
\name{setup_day2}
\alias{setup_day2}
\title{setup all packages for day 2 of the course}
\usage{
setup_day2(newestR = TRUE)
}
\arguments{
\item{newestR}{logical, whether you are using R-3.3.0}
}
\value{
Installs the required packages - day2
}
\description{
setup all packages for day 2 of the course
}
\examples{
\dontrun{
 setup_day2(newestR=TRUE)
}
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2016
}

