% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbioc2016-pkg.R
\docType{data}
\name{surveyrbioc}
\alias{surveyrbioc}
\title{The results of the pre-course survey}
\description{
The results of the pre-course survey, arranged in a \code{data.frame} object.
}
\details{
The following questions were asked to the participants:
\itemize{
 \item{How old are you? Continuous value}
 \item{What is you current academic level Categorical variable; choices: PI/Professor/Group leader or similar,
 Postdoc,PhD student,Master student/else}
 \item{What is your current knowledge level of R? Categorical variable; choices: pro,good,intermediate,poor,none}
 \item{What is your knowledge of programming languages in general? Categorical variable; choices: pro,good,intermediate,poor,none}
 \item{What is your experience level with genomics and RNA-seq data? Categorical variable; choices: pro,good,intermediate,poor,genoWhat?}
 \item{How familiar are you with mogon and parallel computing? Categorical variable; choices: I am a regular user,Once in a while i used it,I know it exists,I heard we had some servers around,Is this supposed to be in the cloud?!}
 \item{What are your expectations from the course? Free text input}
}
}
\author{
Federico Marini, \email{marinif@uni-mainz.de}, 2016
}

